refactor: remove model upgrader tests
This commit removes the units tests for model upgrader. We are actively
trying to migrate this facade off of mongo and the quickest way to do
this to document what we had and make the changes instead of keeping
things going.
